Full stack Project/api/requirements.txt
fastapi
pydantic

# Explicitly including base langchain for chains
langchain
langchain-community
langchain-core
langchain-openai

pinecone
pypdf
python-dotenv
requests
